==> code/Regional Analysis/03.1 - Regional Carbon Emissions and Offset Pricing (Closest Meeting).Rmd <==
---
title: "03.1 - Regional Carbon Emissions and Offset Pricing (Closest Meeting)"
author: "Alec Stashevsky"
date: "1/15/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(httr)
library(jsonlite)
library(readxl)
library(openxlsx)
library(tidyverse)
library(data.table)
library(geosphere)
rm(list = ls()) # Clean R Environment
```


## Import Regional GeoData
```{r Import}
path.in <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/Regional Analysis/APA Regional GeoData.xlsx"

nyc <- read_excel(path.in, sheet = "NYC 2018") %>% setDT()
sf <- read_excel(path.in, sheet = "SF 2019") %>% setDT()
phl <- read_excel(path.in, sheet = "PHL 2020") %>% setDT()
conventions <- read_xlsx(path.in, sheet = "Convention Centers") %>%
  setDT()
```

### Building GoClimate API Queries
```{r Build Query Lists}

# Set Query Parameters
origins.nyc <- nyc$IATA
destination.nyc <- nyc$closestConf

origins.sf <- sf$IATA
destination.sf <- sf$closestConf

origins.phl <- phl$IATA
destination.phl <- phl$closestConf

seat.class <- "economy" # Assume most people will fly economy class

# Initialize list
query.list.nyc <- rep(NA, length(origins.nyc))
query.list.sf <- rep(NA, length(origins.sf))
query.list.phl <- rep(NA, length(origins.phl))

# Generate list for each round-trip query 
## NYC
for (i in 1:length(origins.nyc)){

    query <- paste0("https://api.goclimate.com/v1/flight_footprint?segments[0][origin]=",
                origins.nyc[i],          # Loop through each unique origin airport
                "&segments[0][destination]=",
                destination.nyc[i],         # Nearest airport to convention center
                "&segments[1][origin]=",
                destination.nyc[i],         # 2nd-leg of round trip flight
                "&segments[1][destination]=",
                origins.nyc[i],          # Return home
                "&cabin_class=",
                seat.class,              # Seat class of passenger
                "&currencies[]=USD")     # Price offsets in USD
    
    query.list.nyc[i] <- query
    
}

## SF
for (i in 1:length(origins.sf)){

    query <- paste0("https://api.goclimate.com/v1/flight_footprint?segments[0][origin]=",
                origins.sf[i],          # Loop through each unique origin airport
                "&segments[0][destination]=",
                destination.sf[i],         # Nearest airport to convention center
                "&segments[1][origin]=",
                destination.sf[i],         # 2nd-leg of round trip flight
                "&segments[1][destination]=",
                origins.sf[i],          # Return home
                "&cabin_class=",
                seat.class,              # Seat class of passenger
                "&currencies[]=USD")     # Price offsets in USD
    
    query.list.sf[i] <- query
    
}

## PHL
for (i in 1:length(origins.phl)){

    query <- paste0("https://api.goclimate.com/v1/flight_footprint?segments[0][origin]=",
                origins.phl[i],          # Loop through each unique origin airport
                "&segments[0][destination]=",
                destination.phl[i],         # Nearest airport to convention center
                "&segments[1][origin]=",
                destination.phl[i],         # 2nd-leg of round trip flight
                "&segments[1][destination]=",
                origins.phl[i],          # Return home
                "&cabin_class=",
                seat.class,              # Seat class of passenger
                "&currencies[]=USD")     # Price offsets in USD
    
    query.list.phl[i] <- query
    
}
```


### Query Attendee Data
```{r API Query NYC}
# From GoClimate
api.key <- readRDS("C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GoClimate API/Api_key.rds" )

# Initialize query responses
responses.footprint.nyc <- as.vector(rep(NA, length(query.list.nyc)))
responses.offsets.nyc <- rep(NA, length(query.list.nyc)) 

# Loop through all queries in query.list
for (i in 1:length(query.list.nyc)){
    
    r <- GET(query.list.nyc[i],
             authenticate(api.key, ""))
    
    c <- fromJSON(content(r, as = "text"))
    
    # Ignore null return values from API
    tryCatch({ 
      responses.footprint.nyc[i] <- as.numeric(c$footprint)
      responses.offsets.nyc[i] <- as.numeric(c$offset_prices[1])
    },
    
      error = function(e){
      print(paste0("NULL QUERY ", i))
        
    })
}

# Append to Geodata
nyc.emissions <- cbind(nyc,
                       FootprintcC = responses.footprint.nyc,
                       OffsetcC = responses.offsets.nyc)
```


```{r API Query SF}
# Initialize query responses
responses.footprint.sf <- as.vector(rep(NA, length(query.list.sf)))
responses.offsets.sf <- rep(NA, length(query.list.sf)) 

# Loop through all queries in query.list
for (i in 1:length(query.list.sf)){
    
    r <- GET(query.list.sf[i],
             authenticate(api.key, ""))
    
    c <- fromJSON(content(r, as = "text"))
    
    # Ignore null return values from API
    tryCatch({ 
      responses.footprint.sf[i] <- as.numeric(c$footprint)
      responses.offsets.sf[i] <- as.numeric(c$offset_prices[1])
    },
    
      error = function(e){
      print(paste0("NULL QUERY ", i))
        
    })
}

# Append to Geodata
sf.emissions <- cbind(sf,
                       FootprintcC = responses.footprint.sf,
                       OffsetcC = responses.offsets.sf)
```

```{r API Query PHL}
# Initialize query responses
responses.footprint.phl <- as.vector(rep(NA, length(query.list.phl)))
responses.offsets.phl <- rep(NA, length(query.list.phl)) 

# Loop through all queries in query.list
for (i in 1:length(query.list.phl)){
    
    r <- GET(query.list.phl[i],
             authenticate(api.key, ""))
    
    c <- fromJSON(content(r, as = "text"))
    
    # Ignore null return values from API
    tryCatch({ 
      responses.footprint.phl[i] <- as.numeric(c$footprint)
      responses.offsets.phl[i] <- as.numeric(c$offset_prices[1])
    },
    
      error = function(e){
      print(paste0("NULL QUERY ", i))
        
    })
}

# Append to Geodata
phl.emissions <- cbind(phl,
                       FootprintcC = responses.footprint.phl,
                       OffsetcC = responses.offsets.phl)
```

```{r Export}
## Save Emissions Data
final.sets <- list("NYC 2018" = nyc.emissions,
                 "SF 2019" = sf.emissions,
                 "PHL 2020" = phl.emissions,
                 "Convention Centers" = conventions)

write.xlsx(final.sets,
           file = "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/Regional Analysis/APA Regional Emissions (Closest Conf).xlsx")
```
