==> code/Regional Analysis/04 - Regional Emissions Calculations.Rmd <==
---
title: "04 - Emissions Calculation"
author: "Alec Stashevsky"
date: "12/26/2020"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(readxl)
library(openxlsx)
library(tidyverse)
library(data.table)
rm(list = ls()) # Clean R Environment
```

## Import Data
```{r Import}

path.in <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/Regional Analysis/APA Regional Emissions.xlsx"

nyc <- read_excel(path.in, sheet = "NYC 2018") %>% setDT()
sf <- read_excel(path.in, sheet = "SF 2019") %>% setDT()
phl <- read_excel(path.in, sheet = "PHL 2020") %>% setDT()
conventions <- read_xlsx(path.in, sheet = "Convention Centers") %>%
  setDT()
```

## Prep Calculations
```{r Prep}
# Combine corresponding driving indicators
nyc[, `:=`(Driver = fcase(
  `Regional Destination` == "JFK", drive.JFK,
  `Regional Destination` == "ORD", drive.ORD,
  `Regional Destination` == "PHL", drive.PHL, 
  `Regional Destination` == "YYZ", drive.YYZ,
  `Regional Destination` == "MSY", drive.MSY,
  `Regional Destination` == "SEA", drive.SEA,
  `Regional Destination` == "LAX", drive.LAX),
  
  "Driver Distance" = fcase(
    `Regional Destination` == "JFK", gdist.JFK,
    `Regional Destination` == "ORD", gdist.ORD,
    `Regional Destination` == "PHL", gdist.PHL,
    `Regional Destination` == "YYZ", gdist.YYZ,
    `Regional Destination` == "MSY", gdist.MSY,
    `Regional Destination` == "SEA", gdist.SEA,
    `Regional Destination` == "LAX", gdist.LAX)
  )]

sf[, `:=`(Driver = fcase(
  `Regional Destination` == "JFK", drive.JFK,
  `Regional Destination` == "ORD", drive.ORD,
  `Regional Destination` == "PHL", drive.PHL, 
  `Regional Destination` == "YYZ", drive.YYZ,
  `Regional Destination` == "MSY", drive.MSY,
  `Regional Destination` == "SEA", drive.SEA,
  `Regional Destination` == "LAX", drive.LAX),
  
  "Driver Distance" = fcase(
    `Regional Destination` == "JFK", gdist.JFK,
    `Regional Destination` == "ORD", gdist.ORD,
    `Regional Destination` == "PHL", gdist.PHL,
    `Regional Destination` == "YYZ", gdist.YYZ,
    `Regional Destination` == "MSY", gdist.MSY,
    `Regional Destination` == "SEA", gdist.SEA,
    `Regional Destination` == "LAX", gdist.LAX)
  )]

phl[, `:=`(Driver = fcase(
  `Regional Destination` == "JFK", drive.JFK,
  `Regional Destination` == "ORD", drive.ORD,
  `Regional Destination` == "PHL", drive.PHL, 
  `Regional Destination` == "YYZ", drive.YYZ,
  `Regional Destination` == "MSY", drive.MSY,
  `Regional Destination` == "SEA", drive.SEA,
  `Regional Destination` == "LAX", drive.LAX),
  
  "Driver Distance" = fcase(
    `Regional Destination` == "JFK", gdist.JFK,
    `Regional Destination` == "ORD", gdist.ORD,
    `Regional Destination` == "PHL", gdist.PHL,
    `Regional Destination` == "YYZ", gdist.YYZ,
    `Regional Destination` == "MSY", gdist.MSY,
    `Regional Destination` == "SEA", gdist.SEA,
    `Regional Destination` == "LAX", gdist.LAX)
  )]
```

# Test against main analysis

There are slightly different number of people, import main analysis and check 
observations that are missing.

```{r Import Main}
main.nyc <- read_excel("C:/Users/Alec/Documents/GAP Research/GAP Climate Research/Analysis/APA Emissions All Locations (Analysis).xlsx", sheet = "NYC 2018") %>% setDT()

# Merge to check, frequency
merge.test <- merge(nyc,
  main.nyc[, .(location, lat, lon, Frequency)],
  by = c("location", "lat", "lon"), all.x = TRUE)

sum(merge.test$Frequency.y, na.rm = TRUE)
sum(merge.test$Frequency.x, na.rm = TRUE)

# Replace mismatched frequency w/ main analysis
# ## (since this was edited for some mislabeled data)
# 
# merge.test[Frequency.x != Frequency.y & !is.na(Frequency.y)]$Frequency.x <- merge.test[Frequency.x != Frequency.y & !is.na(Frequency.y)]$Frequency.y
# 
# # Save finalized analysis dataset
# nyc.fix <- merge.test[location %in% main.nyc$location][!is.na(Frequency.x)]

# Rename variables
nyc.fix <- merge.test[, `:=`(
  Frequency = Frequency.x)][
    , Frequency.x := NULL]

# Original column names 
col.order <- c("ATTENDEE_STATUS", "CITY", "STATE", "COUNTRY",	"location",	"lon",
  "lat", "Frequency",	"gdist.PHL", "gdist.JFK", "gdist.ORD", "gdist.MSY",
  "gdist.YYZ", "gdist.LAX", "drive.PHL", "drive.JFK", "drive.ORD", "drive.MSY",
  "drive.YYZ", "drive.LAX",	"ICAO",	"IATA", "Airport Name",	"City",	"Country",
  "Altitude",	"Latitude",	"Longitude", "airport_dist_meters", "gdist.SEA",
  "drive.SEA", "APA.region", "CnumMinDistance", "closestConf",
  "Regional Destination", "Footprint", "Offset", "Driver", "Driver Distance")

# Reorder columns
setcolorder(nyc.fix, col.order)
setnames(nyc.fix, old = "Frequency.y", new = "Frequency Adjusted")
```

```{r Import Main SF}
main.sf <- read_excel("C:/Users/Alec/Documents/GAP Research/GAP Climate Research/Analysis/APA Emissions All Locations (Analysis).xlsx", sheet = "SF 2019") %>% setDT()

# Merge to check, frequency
merge.test.sf <- merge(sf,
  main.sf[, .(location, lat, lon, Frequency)],
  by = c("location", "lat", "lon"), all.x = TRUE)

sum(merge.test.sf$Frequency.y, na.rm = TRUE)
sum(merge.test.sf$Frequency.x, na.rm = TRUE)



# Replace mismatched frequency w/ main analysis 
## (since this was edited for some mislabeled data)

# merge.test.sf[Frequency.x != Frequency.y]$Frequency.x <- merge.test.sf[Frequency.x != Frequency.y]$Frequency.y

# Save finalized analysis dataset remove row of 10 people for "Karachi, VA US"
## Mislabel
# 
# sf.fix <- merge.test.sf[!is.na(Frequency.x)][location != "Karachi, VA, US"]
# sum(sf.fix$Frequency.x)


# Rename variables
sf.fix <- merge.test.sf[, `:=`(
  Frequency = Frequency.x)][
    , Frequency.x := NULL]

# Reorder columns
setcolorder(sf.fix, col.order)
setnames(sf.fix, old = "Frequency.y", new = "Frequency Adjusted")
```


## Export
```{r Export Data}
final.sets <- list("NYC 2018" = nyc.fix,
                   "SF 2019"  = sf.fix,
                   "PHL 2020" = phl,
                   "Convention Centers" = conventions)

write.xlsx(final.sets,
           file = "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/Regional Analysis/APA Regional Emissions (Prepped).xlsx")
```
