==> code/Regional Analysis/04.1 - Regional Emissions Calculations (Closest Meeting).Rmd <==
---
title: "04.1 - Emissions Calculation (Closest Meeting)"
author: "Alec Stashevsky"
date: "1/16/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(readxl)
library(openxlsx)
library(tidyverse)
library(data.table)
rm(list = ls()) # Clean R Environment
```

## Import Data
```{r Import}

path.in <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/Regional Analysis/APA Regional Emissions (Closest Conf).xlsx"

nyc <- read_excel(path.in, sheet = "NYC 2018") %>% setDT()
sf <- read_excel(path.in, sheet = "SF 2019") %>% setDT()
phl <- read_excel(path.in, sheet = "PHL 2020") %>% setDT()
conventions <- read_xlsx(path.in, sheet = "Convention Centers") %>%
  setDT()
```

## Prep Calculations
```{r Prep}
# Combine corresponding driving indicators
nyc[, `:=`(Driver = fcase(
  closestConf == "JFK", drive.JFK,
  closestConf == "ORD", drive.ORD,
  closestConf == "PHL", drive.PHL, 
  closestConf == "YYZ", drive.YYZ,
  closestConf == "MSY", drive.MSY,
  closestConf == "SEA", drive.SEA,
  closestConf == "LAX", drive.LAX),
  
  "Driver Distance" = fcase(
    closestConf == "JFK", gdist.JFK,
    closestConf == "ORD", gdist.ORD,
    closestConf == "PHL", gdist.PHL,
    closestConf == "YYZ", gdist.YYZ,
    closestConf == "MSY", gdist.MSY,
    closestConf == "SEA", gdist.SEA,
    closestConf == "LAX", gdist.LAX)
  )]

sf[, `:=`(Driver = fcase(
  closestConf == "JFK", drive.JFK,
  closestConf == "ORD", drive.ORD,
  closestConf == "PHL", drive.PHL, 
  closestConf == "YYZ", drive.YYZ,
  closestConf == "MSY", drive.MSY,
  closestConf == "SEA", drive.SEA,
  closestConf == "LAX", drive.LAX),
  
  "Driver Distance" = fcase(
    closestConf == "JFK", gdist.JFK,
    closestConf == "ORD", gdist.ORD,
    closestConf == "PHL", gdist.PHL,
    closestConf == "YYZ", gdist.YYZ,
    closestConf == "MSY", gdist.MSY,
    closestConf == "SEA", gdist.SEA,
    closestConf == "LAX", gdist.LAX)
  )]

phl[, `:=`(Driver = fcase(
  closestConf == "JFK", drive.JFK,
  closestConf == "ORD", drive.ORD,
  closestConf == "PHL", drive.PHL, 
  closestConf == "YYZ", drive.YYZ,
  closestConf == "MSY", drive.MSY,
  closestConf == "SEA", drive.SEA,
  closestConf == "LAX", drive.LAX),
  
  "Driver Distance" = fcase(
    closestConf == "JFK", gdist.JFK,
    closestConf == "ORD", gdist.ORD,
    closestConf == "PHL", gdist.PHL,
    closestConf == "YYZ", gdist.YYZ,
    closestConf == "MSY", gdist.MSY,
    closestConf == "SEA", gdist.SEA,
    closestConf == "LAX", gdist.LAX)
  )]
```

# Test against main analysis

Import from original regional analysis which was adjusted according to the main
analysis.

```{r Import Main}
main.nyc <- read_excel("C:/Users/Alec/Documents/GAP Research/GAP Climate Research/Analysis/Regional Analysis/APA Regional Emissions (Analysis).xlsx", sheet = "NYC 2018") %>% setDT()

# Merge to check, frequency
merge.test <- merge(nyc,
  main.nyc[, .(ATTENDEE_STATUS, location, lat, lon, `Frequency Adj`)],
  by = c("ATTENDEE_STATUS", "location", "lat", "lon"), all.x = TRUE)

sum(merge.test$`Frequency Adj`, na.rm = TRUE)
sum(merge.test$Frequency, na.rm = TRUE)

# Fill in 3 missing rows for adjusted freq
merge.test[is.na(`Frequency Adj`)]$`Frequency Adj` <- 1


# Original column names 
col.order <- c("ATTENDEE_STATUS", "CITY", "STATE", "COUNTRY",	"location",	"lon",
  "lat", "Frequency",	"gdist.PHL", "gdist.JFK", "gdist.ORD", "gdist.MSY",
  "gdist.YYZ", "gdist.LAX", "drive.PHL", "drive.JFK", "drive.ORD", "drive.MSY",
  "drive.YYZ", "drive.LAX",	"ICAO",	"IATA", "Airport Name",	"City",	"Country",
  "Altitude",	"Latitude",	"Longitude", "airport_dist_meters", "gdist.SEA",
  "drive.SEA", "APA.region", "CnumMinDistance", "closestConf",
  "Regional Destination", "FootprintcC", "OffsetcC", "Driver", "Driver Distance")

# Reorder columns
setcolorder(merge.test, col.order)
setnames(merge.test, old = "Frequency Adj", new = "Frequency Adjusted")
```

```{r Import Main SF}
main.sf <- read_excel("C:/Users/Alec/Documents/GAP Research/GAP Climate Research/Analysis/Regional Analysis/APA Regional Emissions (Analysis).xlsx", sheet = "SF 2019") %>% setDT()

# Merge to check, frequency
merge.test.sf <- merge(sf,
  main.sf[, .(ATTENDEE_STATUS, location, lat, lon, `Frequency Adj`)],
  by = c("ATTENDEE_STATUS", "location", "lat", "lon"), all.x = TRUE)

sum(merge.test.sf$`Frequency Adj`, na.rm = TRUE)
sum(merge.test.sf$Frequency, na.rm = TRUE)

# Reorder columns
setcolorder(merge.test.sf, col.order)
setnames(merge.test.sf, old = "Frequency Adj", new = "Frequency Adjusted")
```


## Export
```{r Export Data}
final.sets <- list("NYC 2018" = merge.test,
                   "SF 2019"  = merge.test.sf,
                   "PHL 2020" = phl,
                   "Convention Centers" = conventions)

write.xlsx(final.sets,
           file = "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/Regional Analysis/APA Regional Emissions (Closest Conf Prepped).xlsx")
```
